adds owner checking

+++ controllers/contacts.js
@@ -11,7 +11,7 @@ const listContacts = async (req, res) => {
     "-createdAt -updatedAt",
     {
       skip,
-      limit,
+      limit: Number(limit),
     }
   ).populate("owner", "email phone");
   res.json(result);
@@ -19,7 +19,8 @@ const listContacts = async (req, res) => {
 
 const getContactById = async (req, res) => {
   const { id } = req.params;
-  const result = await Contact.findById(id);
+  const { _id: owner } = req.user;
+  const result = await Contact.findById({ _id: id, owner });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -34,7 +35,8 @@ const addContact = async (req, res) => {
 
 const removeContact = async (req, res) => {
   const { id } = req.params;
-  const result = await Contact.findByIdAndRemove(id);
+  const { _id: owner } = req.user;
+  const result = await Contact.findByIdAndRemove({ _id: id, owner });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -45,7 +47,10 @@ const removeContact = async (req, res) => {
 
 const updateById = async (req, res) => {
   const { id } = req.params;
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  const { _id: owner } = req.user;
+  const result = await Contact.findByIdAndUpdate({ _id: id, owner }, req.body, {
+    new: true,
+  });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -54,7 +59,10 @@ const updateById = async (req, res) => {
 
 const updateFavorite = async (req, res) => {
   const { id } = req.params;
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  const { _id: owner } = req.user;
+  const result = await Contact.findByIdAndUpdate({ _id: id, owner }, req.body, {
+    new: true,
+  });
   if (!result) {
     throw HttpError(404, "Not found");
   }
